Refactor: Move setup_helpers.py to sim/input_setup.py for input file generation
