+++ fusion/core/request_generator.py
@@ -1,4 +1,4 @@
-from fusion.helper_scripts.random_helpers import set_seed, get_uniform_rv, get_exponential_rv
+from fusion.utils.random_helpers import set_seed, get_uniform_rv, get_exponential_rv
 
 
 def get_requests(seed: int, engine_props: dict):
