+++ fusion/core/snr_measurements.py
@@ -4,7 +4,7 @@ import numpy as np
 import networkx as nx
 
 from fusion.cli.args.snr_args import SNRProps
-from fusion.helper_scripts.snr_helpers import get_slot_index, get_loaded_files, compute_response
+from fusion.modules.snr.snr_helpers import get_slot_index, get_loaded_files, compute_response
 
 
 # fixme: Only works for seven cores
