+++ fusion/modules/rl/model_manager.py
@@ -8,7 +8,7 @@ from fusion.modules.rl.args.registry_args import ALGORITHM_REGISTRY
 
 from fusion.modules.rl.feat_extrs.constants import CACHE_DIR
 from fusion.modules.rl.feat_extrs.path_gnn_cached import CachedPathGNN
-from fusion.helper_scripts.sim_helpers import parse_yaml_file
+from fusion.sim.sim_helpers import parse_yaml_file
 
 
 def _parse_policy_kwargs(string: str) -> dict:
