+++ fusion/visualization/plot_utils.py
@@ -4,7 +4,7 @@ from statistics import mean
 
 import numpy as np
 
-from fusion.helper_scripts.sim_helpers import dict_to_list, list_to_title, update_matrices
+from fusion.sim.sim_helpers import dict_to_list, list_to_title, update_matrices
 from fusion.cli.args.plot_args import PlotArgs, PlotProps
