+++ tests/test_os_helpers.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest.mock import patch
 
-from fusion.helper_scripts.os_helpers import create_dir
+from fusion.utils.os_helpers import create_dir
 
 
 class TestOSHelpers(unittest.TestCase):
