+++ tests/test_spectrum_helpers.py
@@ -5,7 +5,7 @@ from types import SimpleNamespace
 from unittest.mock import MagicMock
 
 import numpy as np
-from fusion.helper_scripts.spectrum_helpers import SpectrumHelpers
+from fusion.modules.spectrum.spectrum_helpers import SpectrumHelpers
 
 
 class TestSpectrumHelpers(unittest.TestCase):
